Fixed failing integration tests after merge
